refactor: improve imports  #5
