+++ src/features/VisualNovel/routes.tsx
@@ -1,8 +1,6 @@
 import React from 'react';
 import { Navigate, PartialRouteObject } from 'react-router';
-import { OverviewPage } from './pages/OverviewPage/OverviewPage';
-import { ReleasesPage } from './pages/ReleasesPage/ReleasesPage';
-import { VisualNovelPage } from './pages/VisualNovelPage/VisualNovelPage';
+import { OverviewPage, ReleasesPage, VisualNovelPage } from './pages';
 
 /**
  * Creates routes for test module.
